+++ tests/utils/fps_tracker.rs
@@ -10,7 +10,7 @@ use std::time::Duration;
 //-------------------------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------------------------
 
-fn check_expected(tracker: &FPSTracker, expected_avg_delta_secs: f32, expected_fps: u16) -> bool
+fn check_expected(tracker: &FpsTracker, expected_avg_delta_secs: f32, expected_fps: u16) -> bool
 {
     if tracker.average_delta_seconds() != expected_avg_delta_secs
     { println!("avg delta: {}", tracker.average_delta_seconds()); return false; }
@@ -28,14 +28,14 @@ fn check_expected(tracker: &FPSTracker, expected_avg_delta_secs: f32, expected_f
 fn fps_tracker_basic()
 {
     // 1. zero records
-    let mut tracker = FPSTracker::new(0);
+    let mut tracker = FpsTracker::new(0);
 
     assert!(check_expected(&tracker, 0f32, 0u16));
     tracker.update(0.1, Duration::from_secs(1));
     assert!(check_expected(&tracker, 0f32, 0u16));
 
     // 2. one record
-    let mut tracker = FPSTracker::new(1);
+    let mut tracker = FpsTracker::new(1);
 
     assert!(check_expected(&tracker, 0f32, 0u16));
     tracker.update(0.1, Duration::from_secs(1));
@@ -44,7 +44,7 @@ fn fps_tracker_basic()
     assert!(check_expected(&tracker, 0.2, 5u16));
 
     // 3. two records
-    let mut tracker = FPSTracker::new(2);
+    let mut tracker = FpsTracker::new(2);
 
     assert!(check_expected(&tracker, 0f32, 0u16));
     tracker.update(0.05, Duration::from_secs(1));